More component folders added and Async/await func
