+++ post.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <div id="post-wrapper"></div>
-    <script type="module" src="post.js"></script>
+    <script type="module" src="./posts/post.js"></script>
    
 </body>
 </html>
