+++ posts.html
@@ -11,7 +11,7 @@
       <h1>Posts page:</h1>
       <div id="post-wrapper"></div>
  
-    <script type="module" src="posts.js"></script>
+    <script type="module" src="./posts/posts.js"></script>
    
 </body>
 </html>
