+++ posts/posts.js
@@ -1,0 +1,14 @@
+
+import  headerElement from "../header/header.js"
+import{renderAllPosts} from "./postsController.js"
+
+
+headerElement()
+
+
+function init() {
+    renderAllPosts()
+}
+
+init()
+
